+++ applications/services/gui/canvas.c
@@ -171,7 +171,14 @@ void canvas_set_font_direction(Canvas* canvas, CanvasDirection dir) {
 }
 
 void canvas_invert_color(Canvas* canvas) {
-    canvas->fb.draw_color = !canvas->fb.draw_color;
+    if(canvas->fb.draw_color == ColorXOR && momentum_settings.dark_mode) {
+        // XOR is 0x02, invert changes it to 0x00 which is White
+        // Basically like resetting to background color
+        // In Dark Mode, background color is Black instead
+        canvas->fb.draw_color = ColorBlack;
+    } else {
+        canvas->fb.draw_color = !canvas->fb.draw_color;
+    }
 }
 
 void canvas_set_font(Canvas* canvas, Font font) {